feat: add domain stock type
